requirements.txt
# Flask Web Framework
Flask==2.3.3
Flask-CORS==4.0.0

# LangChain and OpenAI
langchain-openai==0.1.7
langchain-community==0.2.4
langchain==0.2.3

# Vector Database
chromadb==0.4.24

# Translation and Language Detection
deep-translator==1.11.4
langdetect==1.0.9

# PDF Processing and OCR
PyMuPDF==1.23.26
pytesseract==0.3.10
Pillow==10.0.1
opencv-python==4.8.1.78

# Utilities
numpy==1.24.4
python-dotenv==1.0.0

# Development
gunicorn==21.2.0
